# docs/api.yaml
openapi: 3.0.1
info:
  title: Goprometric metrics cache API
  version: 1.0.0
servers:
  - url: 'api://127.0.0.1:9111'
paths:
  /count/{ns}/{name}/:
    put:
      operationId: IncCounter
      summary: Increment a counter
      description: Increment a counter
      parameters:
        - in: path
          name: ns
          required: true
          description: Namespace
          example: my_namespace
          schema:
            type: string
        - in: path
          name: name
          required: true
          description: Counter name
          example: my_counter
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CounterRequest'
      responses:
        '200':
          description: Counter was incremented
        '201':
          description: Counter was created and incremented
...
components:
  schemas:
    CounterRequest:
      type: object
      properties:
        labels:
          type: string
          example: "path:/login,status:200"
        help:
          type: string
          example: "Describe metric a bit better"
        step:
          type: float
          example: 1.4
...